+++ node/resolvers/routes/index.ts
@@ -3,6 +3,8 @@ import { method } from '@vtex/api'
 import { getAppId } from '../../constants'
 import { processQueue } from '../../utils/Queue'
 import { getSellerQuote } from './seller/getSellerQuote'
+import { getSellerQuotesPaginated } from './seller/getSellerQuotesPaginated'
+import { saveSellerQuote } from './seller/saveSellerQuote'
 import {
   setSellerResponseMetadata,
   validateSellerRequest,
@@ -33,5 +35,11 @@ export const Routes = {
   },
   getSellerQuote: method({
     GET: createSellerHandlers(getSellerQuote),
+  }),
+  getSellerQuotesPaginated: method({
+    GET: createSellerHandlers(getSellerQuotesPaginated),
+  }),
+  saveSellerQuote: method({
+    POST: createSellerHandlers(saveSellerQuote),
   }),
 }