+++ node/resolvers/utils/sellerQuotesController.ts
@@ -1,4 +1,5 @@
 import { NotFoundError } from '@vtex/api'
+import pLimit from 'p-limit'
 
 import {
   QUOTE_DATA_ENTITY,
@@ -23,21 +24,24 @@ export default class SellerQuotesController {
   private async getSellerQuotes({
     page = 1,
     pageSize = 1,
-    where,
-    sort,
+    where = '',
+    sort = '',
   }: GetQuotesArgs) {
-    return this.ctx.clients.masterdata.searchDocuments<Quote>({
-      dataEntity: QUOTE_DATA_ENTITY,
-      fields: QUOTE_FIELDS,
-      schema: SCHEMA_VERSION,
-      pagination: { page, pageSize },
-      where: `seller=${this.seller} AND (${where})`,
-      sort,
-    })
+    return this.ctx.clients.masterdata.searchDocumentsWithPaginationInfo<Quote>(
+      {
+        dataEntity: QUOTE_DATA_ENTITY,
+        fields: QUOTE_FIELDS,
+        schema: SCHEMA_VERSION,
+        pagination: { page, pageSize },
+        where: `seller=${this.seller} AND (${where})`,
+        sort,
+      }
+    )
   }
 
   private async getSellerQuote(id: string) {
-    const [quote] = await this.getSellerQuotes({ where: `id=${id}` })
+    const { data } = await this.getSellerQuotes({ where: `id=${id}` })
+    const [quote] = data
 
     if (!quote) {
       throw new NotFoundError('seller-quote-not-found')
@@ -55,6 +59,30 @@ export default class SellerQuotesController {
     return { organizationName, costCenterName }
   }
 
+  private async getAllChildrenQuotes(parentQuote: string) {
+    const result: Quote[] = []
+
+    const getQuotes = async (page = 1) => {
+      const quotes = await this.ctx.clients.masterdata.searchDocuments<Quote>({
+        dataEntity: QUOTE_DATA_ENTITY,
+        schema: SCHEMA_VERSION,
+        pagination: { page, pageSize: 100 },
+        fields: ['subtotal'],
+        where: `parentQuote=${parentQuote}`,
+      })
+
+      if (quotes.length) {
+        result.push(...quotes)
+
+        await getQuotes(page + 1)
+      }
+    }
+
+    await getQuotes()
+
+    return result
+  }
+
   public async getFullSellerQuote(id: string) {
     const quote = await this.getSellerQuote(id)
     const { organizationName, costCenterName } = await this.getOrganizationData(
@@ -62,5 +90,82 @@ export default class SellerQuotesController {
     )
 
     return { ...quote, organizationName, costCenterName }
+  }
+
+  public async getSellerQuotesPaginated({
+    page,
+    pageSize,
+    where,
+    sort = 'creationDate DESC',
+  }: {
+    page: number
+    pageSize: number
+    where?: string
+    sort?: string
+  }) {
+    const { data, pagination } = await this.getSellerQuotes({
+      page,
+      pageSize,
+      where,
+      sort,
+    })
+
+    const limit = pLimit(15)
+    const enrichedQuotes = await Promise.all(
+      data.map((quote) =>
+        limit(async () => {
+          const {
+            organizationName,
+            costCenterName,
+          } = await this.getOrganizationData(quote)
+
+          return { ...quote, organizationName, costCenterName }
+        })
+      )
+    )
+
+    return {
+      data: enrichedQuotes,
+      pagination,
+    }
+  }
+
+  public async saveSellerQuote(id: string, fields: Partial<Quote>) {
+    const currentQuote = await this.getSellerQuote(id)
+
+    await this.ctx.clients.masterdata.updatePartialDocument({
+      dataEntity: QUOTE_DATA_ENTITY,
+      schema: SCHEMA_VERSION,
+      fields,
+      id,
+    })
+
+    const { subtotal } = currentQuote
+    const subtotalDelta = (fields?.subtotal ?? subtotal) - subtotal
+    const { parentQuote } = fields
+
+    if (!subtotalDelta || !parentQuote) return
+
+    /**
+     * The seller can update the subtotal of your quote changing items
+     * prices. Therefore, it is necessary to update the subtotal of the
+     * parent quote. The call below is asynchronous as there is no need
+     * to stop the flow because of this operation.
+     */
+    this.getAllChildrenQuotes(parentQuote)
+      .then((childrenQuotes) => {
+        const sumSubtotal = childrenQuotes.reduce(
+          (acc, quote) => acc + quote.subtotal,
+          0
+        )
+
+        this.ctx.clients.masterdata.updatePartialDocument({
+          dataEntity: QUOTE_DATA_ENTITY,
+          schema: SCHEMA_VERSION,
+          fields: { subtotal: sumSubtotal },
+          id: parentQuote,
+        })
+      })
+      .catch(() => null)
   }
 }